Move COmpactDataClause method impls to own cpp file
